rename gradient -> Gradient
