+++ src/index.js
@@ -1,4 +1,4 @@
 import generator from "./generator";
-import Gradient from "./components/gradient";
+import Gradient from "./components/Gradient";
 
 export { generator, Gradient };
