+++ stories/DefaultPresets.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Story, Preview } from "@storybook/addon-docs/blocks";
 import styled from "styled-components";
 
-import Gradient from "../src/components/gradient.js";
+import Gradient from "../src/components/Gradient.js";
 import { listGradients } from "../src/gradients.js";
 
 export const StyledGradient = styled(Gradient)`
